Adding Northstar email status ingestion, and renaming some files.
